refactor limit_order to move out limit_order_txs logic and initialize limit_order dependencies in vm.go
