+++ plugin/evm/limit_order_test.go
@@ -2,8 +2,11 @@ package evm
 
 import (
 	"fmt"
+	"io/ioutil"
 	"testing"
 
+	"github.com/ava-labs/subnet-evm/accounts/abi"
+	"github.com/ava-labs/subnet-evm/plugin/evm/limitorders"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -18,14 +21,22 @@ func TestNewLimitOrderProcesser(t *testing.T) {
 	enabledEthAPIs := []string{"debug"}
 	configJSON := fmt.Sprintf("{\"rpc-tx-fee-cap\": %g,\"eth-apis\": %s}", txFeeCap, fmt.Sprintf("[%q]", enabledEthAPIs[0]))
 	_, vm, _, _ := GenesisVM(t, false, "", configJSON, "")
+	memoryDb := limitorders.NewInMemoryDatabase()
+	jsonBytes, _ := ioutil.ReadFile(orderBookContractFileLocation)
+	orderBookAbi, err := abi.FromSolidityJson(string(jsonBytes))
+	if err != nil {
+		panic(err)
+	}
+	lotp := limitorders.NewLimitOrderTxProcessor(vm.txPool, orderBookAbi, memoryDb, orderBookContractAddress)
 	lop := NewLimitOrderProcesser(
 		vm.ctx,
-		vm.chainConfig,
 		vm.txPool,
 		vm.shutdownChan,
 		&vm.shutdownWg,
 		vm.eth.APIBackend,
 		vm.eth.BlockChain(),
+		memoryDb,
+		lotp,
 	)
 	assert.NotNil(t, lop)
 }
